// frontend/src/models/autenticacion/prematriculaModel.ts
export interface FormData {
  apellidos: string;
  nombres: string;
  tipoDocumento: string;
  numeroDocumento: string;
  fechaNacimiento: string;
  paisNacimiento: string;
  departamentoNacimiento: string;
  municipioNacimiento: string;
  categoriaSisben: string;
  subcategoriaSisben: string;
  direccionResidencia: string;
  telefono: string;
  rutaEscolar: string;
  seguroMedico: string;
  discapacidad: string;
  detalleDiscapacidad: string;
  poblacionDesplazada: string;
  fechaDesplazamiento: string;
  paisResidencia: string;
  departamentoResidencia: string;
  municipioResidencia: string;
  gradoIngreso: string;
  institucionAnterior: string;
  municipioAnterior: string;
  sede: string;
  acudiente1Parentesco: string;
  acudiente1Apellidos: string;
  acudiente1Nombres: string;
  acudiente1CC: string;
  acudiente1Celular: string;
  acudiente1Ocupacion: string;
  acudiente2Parentesco: string;
  acudiente2Apellidos: string;
  acudiente2Nombres: string;
  acudiente2CC: string;
  acudiente2Celular: string;
  acudiente2Ocupacion: string;
}

export const initialForm: FormData = {
  apellidos: '',
  nombres: '',
  tipoDocumento: '',
  numeroDocumento: '',
  fechaNacimiento: '',
  paisNacimiento: 'COLOMBIA',
  departamentoNacimiento: '',
  municipioNacimiento: '',
  categoriaSisben: '',
  subcategoriaSisben: '',
  direccionResidencia: '',
  telefono: '',
  rutaEscolar: '',
  seguroMedico: '',
  discapacidad: '',
  detalleDiscapacidad: '',
  poblacionDesplazada: '',
  fechaDesplazamiento: '',
  paisResidencia: 'COLOMBIA',
  departamentoResidencia: '',
  municipioResidencia: '',
  gradoIngreso: '',
  institucionAnterior: '',
  municipioAnterior: '',
  sede: '',
  acudiente1Parentesco: '',
  acudiente1Apellidos: '',
  acudiente1Nombres: '',
  acudiente1CC: '',
  acudiente1Celular: '',
  acudiente1Ocupacion: '',
  acudiente2Parentesco: '',
  acudiente2Apellidos: '',
  acudiente2Nombres: '',
  acudiente2CC: '',
  acudiente2Celular: '',
  acudiente2Ocupacion: '',
};